Organizando pastas e realizando refatoracões
