+++ src/App.js
@@ -1,25 +1,38 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import React from "react";
-import Login from "./components/Login";
-import Cadastro from "./components/Cadastro";
-import Home from "./components/Home";
+import Login from "./pages/LoginPage/Login";
+import Cadastro from "./pages/CadastroṔage/Cadastro";
+import Home from "./pages/HomePage/Home";
 
 function App() {
 
-  const[habilitado, setHabilitado] = React.useState(false)
-  const [login, setLogin] = React.useState({email:'' , password:''})
-  const [dadosusuario, setDadosUsuario] =  React.useState()
-  
-  return (
+  const [habilitado, setHabilitado] = React.useState(false)
+  const [login, setLogin] = React.useState({ email: '', password: '' })
+  const [dadosusuario, setDadosUsuario] = React.useState()
 
+  return (
     <BrowserRouter>
       <Routes>
-      <Route path='/' element={<Login habilitado={habilitado} setHabilitado={setHabilitado} login={login} setLogin={setLogin} dadosusuario={dadosusuario} setDadosUsuario={setDadosUsuario}></Login>}></Route>
-      <Route path="/cadastro" element={<Cadastro habilitado={habilitado} setHabilitado={setHabilitado}></Cadastro>}></Route>
-      <Route path="/home" element={<Home></Home>}></Route>
+
+        <Route path='/' element={
+          <Login
+            habilitado={habilitado}
+            setHabilitado={setHabilitado}
+            login={login}
+            setLogin={setLogin}
+            dadosusuario={dadosusuario}
+            setDadosUsuario={setDadosUsuario} />} />
+
+        <Route path="/cadastro" element={
+          <Cadastro
+            habilitado={habilitado}
+            setHabilitado={setHabilitado}
+          />} />
+
+        <Route path="/home" element={<Home />} />
+
       </Routes>
     </BrowserRouter>
-
   );
 }
