+++ src/components/Restaurant/RestaurantStyle.js
@@ -1,27 +1,8 @@
-import axios from "axios"
 import styled from "styled-components"
 
-
-
-export default function Restaurant(props){
-    return(
-        <>
-        <RestaurantStyled onClick={() => {
-            
-        }}>
-            {/* Todos estes dados abaixo virão da API. pegar as props que virão e colocar nas devidas informações */}
-            <img src="./assets/img/carterita.png"></img>
-            <div className="description">
-                <h1>Restaurante A</h1>
-                <h2>Comida Brasileira - 50-60 min</h2>
-            </div>
-        </RestaurantStyled>
-        
-     </>
-    )
-}
 // asa
-const RestaurantStyled = styled.div`
+export const RestaurantStyled = styled.div`
+
     margin-bottom: 10px;
     position:relative;
     display:flex;
@@ -37,9 +18,11 @@ const RestaurantStyled = styled.div`
     font-weight: 700;
     line-height: 18px;
     color: #ffae00;;
+
     img{
         height: 100%;
     }
+
     .description { 
         flex-direction: column;
         justify-content: flex-start;
@@ -57,7 +40,6 @@ const RestaurantStyled = styled.div`
             font-size: 15px;
         }
     }
-
 `
